backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
pandas
matplotlib
yfinance
pandas-datareader
prophet
statsmodels
scikit-learn
tensorflow
numpy
textblob
newsapi-python
transformers
torch 